backend/pyproject.toml
[tool.poetry]
name = "expense-manager-backend"
version = "0.1.0"
description = "Backend API for Expense Manager AI"
authors = ["LSantanche"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.46"
alembic = "^1.18.3"
psycopg = {extras = ["binary"], version = "^3.3.2"}
streamlit = "^1.54.0"
httpx = "^0.28.1"
python-multipart = "^0.0.22"
pytesseract = "^0.3.13"
pillow = "^12.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
